requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
selenium==4.10.0
webdriver-manager==3.8.6
pytesseract==0.3.10
pillow==10.0.0
python-multipart==0.0.6